// src/modules/items/weapons/types/ids.type.ts
export type WeaponId =
  // Handguns
  | "9mm-semi-automatic-handgun"
  | "9mm-semi-automatic-handgun+1"
  | "9mm-semi-automatic-handgun+2"
  | "9mm-semi-automatic-handgun+3"
  | "9mm-semi-automatic-handgun+4"
  | "9mm-semi-automatic-handgun+5"
  | "desert-eagle"
  | "desert-eagle+1"
  | "desert-eagle+2"
  | "desert-eagle+3"
  | "desert-eagle+4"
  | "desert-eagle+5"

  // Shotguns
  | "remington-870-pump-action-shotgun"
  | "remington-870-pump-action-shotgun+1"
  | "remington-870-pump-action-shotgun+2"
  | "remington-870-pump-action-shotgun+3"
  | "remington-870-pump-action-shotgun+4"
  | "remington-870-pump-action-shotgun+5"
  | "spas-12"
  | "spas-12+1"
  | "spas-12+2"
  | "spas-12+3"
  | "spas-12+4"
  | "spas-12+5"

  // Submachine Guns (SMG)
  | "uzi"
  | "uzi+1"
  | "uzi+2"
  | "uzi+3"
  | "uzi+4"
  | "uzi+5"
  | "mp5"
  | "mp5+1"
  | "mp5+2"
  | "mp5+3"
  | "mp5+4"
  | "mp5+5";